A lot happened here. Code has been implemented for the loading ramp mechanism.
Button 10 is now being used for this functionality. All code pertaining to the funnelator override button (formerly button 10) has been commented out. It will be a momentary button.

A WsServo class was created, as was a WsLoadingRamp subsystem. Furthermore, a setAngle method did not exist in the Servo class of WsSimulation, so one was added.

This has all been tested on the simulator. However, in the future config values could be added for the two servo values. Also, this should be tested on the actual robot.